Rename `ObjectEmbed` to `Serializer`. Per #11, the `ObjectEmbed` middleware is more aptly named `Serializer`, since object embedding has no actual effect until an Object is serialized, meaning the middleware would be serializing objects before doing any embedding.
Classifying the middleware as `Serializer` covers this process semantically, and generalizes the middleware, allowing it to do more than just embed objects (though, it probably won't any time soon).
